Generate the appropriate citation for your data

Generate the appropriate citation for your data

neon_citation(product = NULL, download_date = Sys.Date(), dir = neon_dir())

Arguments

product
A NEON productCode or list of product codes, see examples.

download_date
Date of download to be included in citation. default is today's date, see details.

dir
Location where files should be downloaded. By default will use the appropriate applications directory for your system (see tools::R_user_dir()). This default also be configured by setting the environmental variable NEONSTORE_HOME, see Sys.setenv or Renviron.

Value

returns a utils::bibentry object, which can be used as text or formatted for bibtex.

Details

Note that the neon_download() does not record download date for each file. Citing a single product download date is after all rather meaningless, as parts of a products may not have all been downloaded on different dates. Indeed, neon_download() is designed in precisely this way, to allow easy updating of downloads without re-downloading older data.

References

https://www.neonscience.org/data-samples/data-policies-citation

Examples

neon_citation("DP1.10003.001")

#> National Ecological Observatory Network (2021). “Data Products:
#> NEON.DP1.10003.001 . Provisional data downloaded from
#> http://data.neonscience.org on 02 Mar 2021.”

## or the citation for all products in store:
neon_citation()

#> National Ecological Observatory Network (2021). “Provisional data
#> downloaded from http://data.neonscience.org on 02 Mar 2021.”

## as bibtex
format(neon_citation("DP1.10003.001"), "bibtex")

#> [1] "@Misc{,\n  author = {{National Ecological Observatory Network}},\n  year = {2021},\n  title = {Data Products: NEON.DP1.10003.001 . Provisional data downloaded from http://data.neonscience.org on 02 Mar 2021},\n  publisher = {Battelle},\n  location = {Boulder, CO, USA},\n}"