export local neon store as a zip archive

Export all or select files from your neon store as a zip archive. This can be useful if you want to bypass accessing the API, such as for archiving the files required for your analysis so that they can be re-created by other users without an API key, or without waiting for the individual download, or any other tiem you want to share or distribute your local store.

neon_export(
  archive = paste(Sys.Date(), "neonstore.zip", sep = "-"),
  product = NA,
  table = NA,
  site = NA,
  start_date = NA,
  end_date = NA,
  type = NA,
  ext = NA,
  timestamp = NA,
  hash = NULL,
  dir = neon_dir()
)

Arguments

archive
path to the zip archive to be created.#'

product
A NEON productCode or list of product codes, see examples.

table
Include only files matching this table name (or regex pattern). (optional).

site
4-letter site code(s) to filter on. Leave as NA to search all.

start_date
Download only files as recent as (YYYY-MM-DD). Leave as NA to download up to the most recent available data.

end_date
Download only files up to end_date (YYYY-MM-DD). Leave as NA to download all prior data.

type
Should we prefer the basic or expanded version of this product? Note that not all products have expanded formats.

ext
only match files with this file extension(s)

timestamp
only match timestamps prior this. See details in neon_index(). Should be a datetime POSIXct object (or coerce-able string)

hash
name of a hashing algorithm to check file integrity. Can be "md5", "sha1", or "sha256" currently; or set to NULL (default) to skip hash computation.

dir
Location where files should be downloaded. By default will use the appropriate applications directory for your system (see tools::R_user_dir()). This default also be configured by setting the environmental variable NEONSTORE_HOME, see Sys.setenv or Renviron.

Value

table of selected files and metadata, from neon_index(), invisibly.

See also

Examples

archive <- tempfile()
dir <- tempdir()
neon_export(archive, dir = dir)