NEON filename parser

Parse filenames into their component metadata. See details for definition of each metadata field, or consult the NEON documentation linked below. https://data.neonscience.org/file-naming-conventions

neon_filename_parser(x)

Arguments

x

vector of NEON filenames

Value

a data frame in which filenames have been split into metadata components. Column names indicate the metadata field code, see details section for complete descriptions.

Details

Metadata components:

NEON A four-character alphanumeric code, denoting the organizational origin of the data product and identifying the product as operational; data collected as part of a special data collection exercise are designated by a separate, unique alphanumeric code created by the PI.

DOM A three-character alphanumeric code, referring to the domain of data acquisition (D01 - D20).

SITE A four-character alphanumeric code, referring to the site of data acquisition; all sites are designated by a standardized four-character alphabetic code.

DPL A three-character alphanumeric code, referring to data product processing level.

PRNUM A five-character numeric code, referring to the data product number (see the Data Product Catalog at http://data.neonscience.org/data-product-catalog).

REV A three-digit designation, referring to the revision number of the data product. The REV value is incremented by 1 each time a major change is made in instrumentation, data collection protocol, or data processing such that data from the preceding revision is not directly comparable to the new.

HOR A three-character alphanumeric code for Spatial Index #1. Refers to measurement locations within one horizontal plane. For example, if five surface measurements were taken, one at each of the five soil array plots, the number in the HOR field would range from 001-005.

VER A three-character alphanumeric code for Spatial Index #2. Refers to measurement locations within one vertical plane. For example, if eight temperature measurements are collected, one at each tower vertical level, the number in the VER field would range from 010-080.

TMI A three-character alphanumeric code for the Temporal Index. Refers to the temporal representation, averaging period, or coverage of the data product (e.g., minute, hour, month, year, sub-hourly, day, lunar month, single instance, seasonal, annual, multi-annual). 000 = native resolution, 001 = native resolution or 1 minute, 002 = 2 minute, 005 = 5 minute, 015 = 15 minute, 030 = 30 minute, 060 = 60 minutes or 1 hour, 101-103 = native resolution of replicate sensor 1, 2, and 3 respectively, 999 = Sensor conducts measurements at varied interval depending on air mass.

DESC An abbreviated description of the data file or table.

YYYY-MM Represents the year and month of the data in the file.

PKGTYPE The type of data package downloaded. Options are 'basic', representing the basic download package, or 'expanded', representing the expanded download package (see more information below).

GENTIME The date-time stamp when the file was generated, in UTC. The format of the date-time stamp is YYYYMMDDTHHmmSSZ.

AOP Products Only (Airborne Observation Platform):

FLHTDATE Date of flight, YYYYMMDD

FLIGHTSTRT Start time of flight, YYYYMMDDHH

FLHTSTRT Start time of flight, YYMMDDHH

IMAGEDATETIME Date and time of image capture, YYYYMMDDHHmmSS

CCCCCC Digital camera serial number

NNNN Sequential number for indexing files

NNN Planned flightline number

R Repeat number

FFFFFF Numeric code for an individual flightline

EEEEEE UTM easting of lower left corner

NNNNNNN UTM northing of lower left corner

References

https://data.neonscience.org/file-naming-conventions