read in neon tabular data

read in neon tabular data

neon_read(
  table = NA,
  product = NA,
  site = NA,
  start_date = NA,
  end_date = NA,
  ext = NA,
  timestamp = NA,
  release = NA,
  dir = neon_dir(),
  files = NULL,
  sensor_metadata = TRUE,
  keep_filename = FALSE,
  altrep = FALSE,
  ...
)

Arguments

table
the name of a downloaded NEON table in the store, see neon_index

product
A NEON productCode or list of product codes, see examples.

site
4-letter site code(s) to filter on. Leave as NA to search all.

start_date
Download only files as recent as (YYYY-MM-DD). Leave as NA to download up to the most recent available data.

end_date
Download only files up to end_date (YYYY-MM-DD). Leave as NA to download all prior data.

ext
only match files with this file extension(s)

timestamp
only match timestamps prior this. See details in neon_index(). Should be a datetime POSIXct object (or coerce-able string)

release
Select only data files associated with a particular release tag, see https://www.neonscience.org/data-samples/data-management/data-revisions-releases, e.g. "RELEASE-2021". Releases are associated with a specific DOI and the promise that files associated with a particular release will not change.

dir
Location where files should be downloaded. By default will use the appropriate applications directory for your system (see tools::R_user_dir()). This default also be configured by setting the environmental variable NEONSTORE_HOME, see Sys.setenv or Renviron.

files
optionally, specify a vector of file paths directly (e.g. as provided from neon_index) and specify table argument as NULL.

sensor_metadata
logical, default TRUE. Should we add metadata fields from file names of sensor data into the table? Adds DomainID, SiteID, horizontalPosition, verticalPosition, and publicationDate. Results in slower parsing.

keep_filename
Should we include a column indicating the original file name for each row? Can be a useful source of additional metadata that NEON may omit from the raw files (i.e. siteID), but will also result in slower parsing. Default FALSE.

altrep
enable or disable altrep. Logical, default FALSE. Setting to TRUE can speed up reading, but may cause vroom::vroom to throw mapping error: Too many open files.

...
additional arguments to vroom::vroom, can usually be omitted.

Details

NEON's tabular data files are separated out into separate .csv files for each site for each month of sampling. In principle, each file has identical columns. vroom::vroom can read in a data table that has been sharded into many files like this much much faster than other parsers can read in each table iteratively, (and thus can greatly out-perform the 'stacking" methods in neonUtilities).

When reading in very large numbers of files, it may be helpful to set altrep = FALSE to opt out of vroom's fast altrep mechanism, which can cause neon_read() to fail when stacking thousands of files.

Unfortunately, not all datasets are entirely consistent in their use of columns. neon_read works around this by parsing such tables in groups of matching schema, which is still reasonably fast.

NEON sensor data products currently do not include important metadata columns containing DomainID, SiteID, horizontalPosition, verticalPosition, and publicationDate in the data files themselves, but only encode this in the in the raw file names. All though these values are shared across a raw data file, this information is lost when stacking the tables unless explicit columns are added to the data. This requires us to parse the files one-by-one, which is much slower. By default this information is added to the table, altering the stacked table schema from that of the raw table. Disable this behavior by setting sensor_metadata = FALSE. Future NEON sensor data products may start including this information in the raw data files, as is already the case for observational data.

Examples

neon_read("brd_countdata-expanded")

#> Warning: number of columns of result is not a multiple of vector length (arg 1)

#>   Some raw data files have changed.
#>   Using only most updated file to avoid duplicates.
#>   see ?neonstore::show_deprecated_data() for details.

#> # A tibble: 4,654 x 24
#>    uid                namedLocation     domainID siteID plotID plotType  pointID
#>    <chr>              <chr>             <chr>    <chr>  <chr>  <chr>     <chr>  
#>  1 bb46d0cc-0102-449… YELL_009.birdGri… D12      YELL   YELL_… distribu… C3     
#>  2 f252dd4d-6684-4e3… YELL_009.birdGri… D12      YELL   YELL_… distribu… C3     
#>  3 b2d4cb65-38ec-42a… YELL_009.birdGri… D12      YELL   YELL_… distribu… C3     
#>  4 2882f646-59ce-47f… YELL_009.birdGri… D12      YELL   YELL_… distribu… C3     
#>  5 2aa0d21f-fdc4-496… YELL_009.birdGri… D12      YELL   YELL_… distribu… C3     
#>  6 61d465df-d551-48a… YELL_009.birdGri… D12      YELL   YELL_… distribu… C3     
#>  7 ea1ffb98-fceb-486… YELL_009.birdGri… D12      YELL   YELL_… distribu… C3     
#>  8 4d3f6161-26bc-401… YELL_009.birdGri… D12      YELL   YELL_… distribu… C3     
#>  9 a93bf746-0df8-462… YELL_009.birdGri… D12      YELL   YELL_… distribu… C3     
#> 10 643c9113-2906-4d1… YELL_009.birdGri… D12      YELL   YELL_… distribu… C3     
#> # … with 4,644 more rows, and 17 more variables: startDate <dttm>,
#> #   eventID <chr>, pointCountMinute <dbl>, targetTaxaPresent <chr>,
#> #   taxonID <chr>, scientificName <chr>, taxonRank <chr>, vernacularName <chr>,
#> #   family <chr>, nativeStatusCode <chr>, observerDistance <dbl>,
#> #   detectionMethod <chr>, visualConfirmation <chr>, sexOrAge <chr>,
#> #   clusterSize <dbl>, clusterCode <chr>, identifiedBy <chr>

## Sensor inputs will add metadata columns by default
neon_read("waq_instantaneous", site = c("CRAM","SUGG"))

#> Warning: number of columns of result is not a multiple of vector length (arg 1)

#> # A tibble: 65,666 x 34
#>    startDateTime       endDateTime         sensorDepth sensorDepthExpUncert
#>    <dttm>              <dttm>                    <dbl> <chr>               
#>  1 2019-04-28 00:00:00 2019-04-28 00:00:00          NA NA                  
#>  2 2019-04-28 00:05:00 2019-04-28 00:05:00          NA NA                  
#>  3 2019-04-28 00:10:00 2019-04-28 00:10:00          NA NA                  
#>  4 2019-04-28 00:15:00 2019-04-28 00:15:00          NA NA                  
#>  5 2019-04-28 00:20:00 2019-04-28 00:20:00          NA NA                  
#>  6 2019-04-28 00:25:00 2019-04-28 00:25:00          NA NA                  
#>  7 2019-04-28 00:30:00 2019-04-28 00:30:00          NA NA                  
#>  8 2019-04-28 00:35:00 2019-04-28 00:35:00          NA NA                  
#>  9 2019-04-28 00:40:00 2019-04-28 00:40:00          NA NA                  
#> 10 2019-04-28 00:45:00 2019-04-28 00:45:00          NA NA                  
#> # … with 65,656 more rows, and 30 more variables: sensorDepthFinalQF <chr>,
#> #   specificConductance <dbl>, specificConductanceExpUncert <dbl>,
#> #   specificCondFinalQF <dbl>, dissolvedOxygen <dbl>,
#> #   dissolvedOxygenExpUncert <dbl>, dissolvedOxygenFinalQF <dbl>,
#> #   dissolvedOxygenSaturation <dbl>, dissolvedOxygenSatExpUncert <dbl>,
#> #   dissolvedOxygenSatFinalQF <dbl>, pH <dbl>, pHExpUncert <dbl>,
#> #   pHFinalQF <dbl>, chlorophyll <dbl>, chlorophyllExpUncert <dbl>,
#> #   chlorophyllFinalQF <dbl>, turbidity <dbl>, turbidityExpUncert <dbl>,
#> #   turbidityFinalQF <dbl>, fDOM <chr>, rawCalibratedfDOM <dbl>,
#> #   fDOMExpUncert <chr>, fDOMFinalQF <dbl>, buoyNAFlag <dbl>, file <chr>,
#> #   domainID <chr>, siteID <chr>, horizontalPosition <chr>,
#> #   verticalPosition <chr>, publicationDate <dttm>