neon_remote select a table from the remote connection

neon_remote

select a table from the remote connection

neon_remote(table = "", product = "", type = "", db = neon_remote_db())

Arguments

table

table name (pattern match regex)

product

product code

type

basic or extended (if necessary to distinguish)

db

a neon_remote_db connection. If not provided, one will be created, but it is faster to pass this on for re-use in multiple neon_remote calls.

Value

a arrow::FileSystemDataset object, or a named list of such objects if multiple matches are found. This table is not downloaded but remains on the remote storage location, but can be filtered with dplyr functions like filter and select, and can also be grouped and summarised, all without ever downloading the whole table. Use dplyr::collect() to download the (possibly filtered) table into and pull into memory.