import neon data into a local database

import neon data into a local database

neon_store(
  table = NA,
  product = NA,
  type = NA,
  dir = neon_dir(),
  db = neon_db(neon_db_dir(), read_only = FALSE),
  n = 500L,
  quiet = FALSE,
  ...
)

Arguments

table

Include only files matching this table name (or regex pattern). (optional).

product

A NEON productCode or list of product codes, see examples.

type

Should we prefer the basic or expanded version of this product? Note that not all products have expanded formats.

dir

Location where files should be downloaded. By default will use the appropriate applications directory for your system (see tools::R_user_dir()). This default also be configured by setting the environmental variable NEONSTORE_HOME, see Sys.setenv or Renviron.

db

A connection to a write-able relational database backend, see neon_db().

n

number of files that should be read per iteration

quiet

show progress?

...

Arguments passed on to neon_read

sensor_metadata

logical, default TRUE. Should we add metadata fields from file names of sensor data into the table? Adds DomainID, SiteID, horizontalPosition, verticalPosition, and publicationDate. Results in slower parsing.

keep_filename

Should we include a column indicating the original file name for each row? Can be a useful source of additional metadata that NEON may omit from the raw files (i.e. siteID), but will also result in slower parsing. Default FALSE.

altrep

enable or disable altrep. Logical, default FALSE. Setting to TRUE can speed up reading, but may cause vroom::vroom to throw mapping error: Too many open files.

files

optionally, specify a vector of file paths directly (e.g. as provided from neon_index) and specify table argument as NULL.

ext

only match files with this file extension(s)

timestamp

only match timestamps prior this. See details in neon_index(). Should be a datetime POSIXct object (or coerce-able string)

start_date

Download only files as recent as (YYYY-MM-DD). Leave as NA to download up to the most recent available data.

end_date

Download only files up to end_date (YYYY-MM-DD). Leave as NA to download all prior data.

site

4-letter site code(s) to filter on. Leave as NA to search all.

release

Select only data files associated with a particular release tag, see https://www.neonscience.org/data-samples/data-management/data-revisions-releases, e.g. "RELEASE-2021". Releases are associated with a specific DOI and the promise that files associated with a particular release will not change.

Value

the index of files read in (invisibly)