Default directory for persistent NEON file store

Use neon_dir() to view or access the currently active local store. By default, neon_download() downloads files into the neon_dir(), which uses an appropriate application directory for your operating system, see tools::R_user_dir(). This location can be overridden by setting the environmental variable NEONSTORE_HOME. neonstore functions (e.g. neon_index(), and neon_read()) look for files in the neon_dir() directory by default. (All functions can also take a one-off argument to dir in the function call in place of the calling neon_dir() to access the default.

neon_dir()

Value

the active neonstore directory.

Examples

neon_dir()
#> [1] "/home/runner/.local/share/R/neonstore"

## Override with an environmental variable:
Sys.setenv(NEONSTORE_HOME = tempdir())
neon_dir()
#> [1] "/tmp/RtmpGips2H"
## Unset
Sys.unsetenv("NEONSTORE_HOME")