Return a neon table from the database

Return a neon table from the database

neon_table(table, site = NA, db = neon_db())

Arguments

table
the name of a downloaded NEON table in the store, see neon_index

site
4-letter site code(s) to filter on. Leave as NA to search all.

db
a connection to the database, see [neon_db()].

Details

We cannot filter on start_date or end_date since these come only from the filename metadata and are only added to instrument tables, not observation tables etc.